Implemented correlation with pertContr
pertContr is defined as the stressed - unstressed difference in the amount
of intensity perturbation, in dB, averaged over all trials in the pert
phase.

The detailed definition was created by Kevin:
(MicINstressed - MicINunstressed) - (FeedBackstressed -
FeedBackunstressed)

Note that this value should be positive for the DN perturbation and
negative for the UP perturbation, if the perturbation were working as
intended and focused on the stressed word.

Usage example:
intShift_analysis_1(1, [1, 2], 'cpa', 'corrPert')

Also added the new "loadCache" option.